+++ lib/src/api/file_system_handle.dart
@@ -1,20 +1,30 @@
+import 'dart:js_util';
+
 import 'package:file_system_access_api/src/api/file_system_kind.dart';
 import 'package:file_system_access_api/src/api/permissions.dart';
+import 'package:file_system_access_api/src/interop/file_system_options.dart';
+import 'package:js/js.dart';
+
+@JS()
+@staticInterop
+abstract class FileSystemHandle {}
 
 /// Represents a file or directory entry. Multiple handles can represent the same entry. For the most part you do not
 /// work with [FileSystemHandle] directly but rather its child interfaces [FileSystemFileHandle] and
 /// [FileSystemDirectoryHandle].
 ///
 /// Example and more on [MDN Web Docs](https://developer.mozilla.org/docs/Web/API/FileSystemHandle)
-abstract class FileSystemHandle {
+extension JSFileSystemHandle on FileSystemHandle {
   /// Returns the type of entry. [file] if the associated entry is a file or [directory].
-  FileSystemKind get kind;
+  FileSystemKind get kind => FileSystemKind.values.byName(getProperty(this, "kind"));
 
   /// Returns the name of the associated entry (extension included).
-  String get name;
+  external String get name;
 
   /// Compares two handles to see if the associated entries (either a file or directory) match.
-  Future<bool> isSameEntry(FileSystemHandle other);
+  Future<bool> isSameEntry(FileSystemHandle other) {
+    return promiseToFuture(callMethod(this, "isSameEntry", [other]));
+  }
 
   /// Queries the current permission state of the current handle.
   ///
@@ -26,16 +36,20 @@ abstract class FileSystemHandle {
   /// can be done. If this returns [denied] any operations will reject. Usually handles returned by the local file
   /// system handle factories will initially return [granted] for their read permission state. However, other than
   /// through the user revoking permission, a handle retrieved from IndexedDB is also likely to return [prompt].
-  Future<PermissionState> queryPermission({required PermissionMode mode});
+  Future<PermissionState> queryPermission({required PermissionMode mode}) async {
+    final options = [FileSystemHandlePermissionDescriptor(mode: mode.name)];
+    final permission = await promiseToFuture(callMethod(this, "queryPermission", options));
+
+    return PermissionState.values.byName(permission);
+  }
 
   /// Requests [read] or [readwrite] permissions for the file handle.
   ///
   /// Returns a [PermissionState] value which is one of [granted], [denied] or [prompt].
-  Future<PermissionState> requestPermission({required PermissionMode mode});
+  Future<PermissionState> requestPermission({required PermissionMode mode}) async {
+    final options = [FileSystemHandlePermissionDescriptor(mode: mode.name)];
+    final permission = await promiseToFuture(callMethod(this, "requestPermission", options));
 
-  /// Returns underlying JavaScript object to store handle in IndexedDB.
-  ///
-  /// **Note:** this feature is not working as-this. See this
-  /// [issue #50621](https://github.com/dart-lang/sdk/issues/50621) and in `example/` folder for a workaround.
-  dynamic toNative();
+    return PermissionState.values.byName(permission);
+  }
 }